skeinforge-35,39 dimension.py: Allow "absolute E" to grow to 100000.00
Previously every M101/M103 included a G92 E0 in absolute E mode.  According
to http://reprap.org/wiki/GCodes (and by firmware inspection) this is an
unbuffered command which flushes the firmware's command buffer.  This change
allows absolute E to grow to a value which retains 0.01 precision in the
firmware's 32-bit float before a reset.